Fix the mocha setup file to resuse app configuration
